Cache KColorScheme instances
Summary:
This patch catches the few different KColorScheme we use so that every
widget can access the same instead of computing it separately.
As is, the different KColorScheme were initialised really often for every
widget that specifies a colorSet.

Test Plan: Used it with Discover and Kirigami Gallery

Reviewers: #frameworks, mart

Reviewed By: mart

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D9768