Fake mousehover whilst combobox is open
Summary:
Under QWidgets the combobox menu steals the mouse focus, including under
wayland. Because of this the root control remains visually hovered as
you move the mouse through the window.

QtQuick combo box is different and that doesn't apply.

This patch is arguably a hack, but it provides the same QStyleOption
hints for widgets and QQC2 which is our goal.

Test Plan: Lots of debug in the QStyle

Reviewers: broulik

Reviewed By: broulik

Subscribers: broulik, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D18285