Use PointingHand when hovering links in Label
This patch simply adds the behaviour usually expected when hovering
above links of changing the mouse cursor to a pointing hand rather
than an arrow. Based on a short discussion in the Discover chat,
brought on by D19118, this patch introduces the fix from there in a
more central location.

Differential Revision: https://phabricator.kde.org/D19179