Update size hint when font changes
Ensures controls update their size when font changes.
The Labels inside would but the control around them not,
leading to broken rendering until causing it to update by interacting with it.
It has to be queued for some reason or else by the time the signal is invoked,
the style hasn't caught up yet.

Differential Revision: https://phabricator.kde.org/D25983