[TabBar] Remove frame
Summary:
In the Window Decorations and Audio KCMs we use QQC2 TabBars with a frame.

The frame already draws a top border but then the tabbar draws another border, resulting in an ugly, dark, out of place line.

This patch removes the frame from the QQC2 TabBar style to fix the issue.

Test Plan:
Before:
{F7787544}

After:
{F7787554}

Reviewers: #plasma, #vdg, ngraham, broulik, mart

Reviewed By: #plasma, #vdg, ngraham, mart

Subscribers: mart, ahiemstra, broulik, ngraham, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D25584