Detect QQC2 version at build time with actual detection.
Summary:
Whenever qqc release a new version the version need to be updated.
Otherwise the feature is not available to user. This change tries
to use real qml code to detect the version by creating object until
success, force to use a minimal qpa platform when running it.

Test Plan:
Run the cmake under a env without display server and version is detected
correctly.

Reviewers: #plasma, rizzitello, davidedmundson, apol, mart

Reviewed By: #plasma, davidedmundson

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D28859