Set editable to true by default for SpinBox
The number of places where we want to have editable spinboxes far
exceeds the number of places where we want non-editable spinboxes. So
make the default editable, it can always be overridden by users of
SpinBox.

This also makes it behave similar to the widgets spinbox, that
apparently does not even have a non-editable mode.