ComboBox.contentItem must be a QQuickTextInput to fix autocomplete
See QQuickComboBoxPrivate::updateEditText()
https://github.com/qt/qtquickcontrols2/blob/dev/src/quicktemplates2/qquickcombobox.cpp#L426

By removing the MouseArea previously in the contentItem, we lose the
ability to press+hold+drag to select a drop down item.

A MouseArea has been added to the background component to capture
wheel events in a similar pattern to PlasmaComponents3.ComboBox.

The ComboBox previously closed the popup when releasing the click to
open the dropdown when editable=true.

BUG: 425865