Fix toolbuttons with menus using newer Breeze
Now that Breeze has a toolbutton-with-instant-menu style, clean up the
things needed for Kirigami to use it. This involves removing some extra
spacing, making sure we do not set AutoRaise for non-flat buttons and
exposing the menu property as an actual property on the background. The
last one is done because setting a property on the properties map breaks
all bindings.