Don't use parent height/width for implicit ToolSeparator sizing
I realized this was pretty much all wrong, so I rewrote the whole thing with a structure inspired by the Default QQC2 theme.

Padding property was removed since it was originally added to give the
separator back the padding that it lost by having the same height as the
parent.

BUG: 425949