Draw ComboBox text in contentItem
We were doing this for the editable case, we may as well do it for both.

This means a user of the desktop style can override background and still
get text.

Technically breaks the idea of some theme doing something crazy for
`CE_ComboBoxLabel` but ultimately the two goals of what this is doing
and what's wanted in the bug report are at odds with each other.

Fixing indicators is not addressed in that patch.

CCBUG: 390791