[Label] Remove unnecessary code
Palette is automatically set to the system palette.

activeFocusOnTab is false by default.

font is automatically set to the system or parent's font.

textColor is already changed to 50% opacity when the Label is not
enabled because of the QPalette::Disabled ColorGroup. No need to reduce
opacity twice.

StaticText Accessible role and name are set automatically for QQC2
Labels, so no need to explicitly set the same exact role and name.