Draw icon in QQStyleItem
Checkbox already knew how to do this, but it was doing so internally to
itself. Let's move that logic to CheckDelegate.qml so that everything
that uses it gains the same feature. This functionality can be optionally
disabled in case a control using it already handles icon-drawing by itself.

CCBUG: 433433