Bind default font of Label
There is no propagation of QGuiApplication::font changes for Text or
Label. This means that changing the default font in systemsettings
has no effect on Text elements in QML. For Label at least we can
work around this by explicitely binding the font to the defaultFont
of the Kirigami Theme which we correctly update.
BUG:438436
FIXED-IN:5.84