Don't change MenuItem's arrow's color when selected
This makes it more consistent with the new direction we're going
with Breeze, the text color and arrow shouldn't change it's colors
when selected.

As Noah Davis said: "we should really do is change how the selected
effect  is handled. I think that's in KIconThemes. We only support
Breeze in this component right now, so removing that line should work
fine. When we support more styles, then we can add it back."

BUG: 443453