Thicken menus on transient touch as well as being in Tablet Mode
In general it's not considered great UX for visible UI elements to move
around or resize in response to touch input, but this does not apply to
menus, because they aren't visible until a touch is used to make them
visible. Therefore we can make them bigger on any kind of transient
touch input, not only when we're in Tablet Mode. This improves menu
usability of QQC2 menus when using a non-tablet touchscreen.

The condition is changed rather than added because
`hasTransientTouchInput` internally takes into account the value of
`tabletMode`.