Replace connect to self in KQuickStyleItem with direct method calls
There is no need to use `connect(this, ... , this, ...)` most of the
time, the needed code can just be called directly. This makes the code
easier to read, less error prone (there were several duplicate connects)
and reduces the memory usage for each instance of KQuickStyleItem.

From testing, this reduces memory usage of 10000 instances of
KQuickStyleItem by around 25 MiB.