Fix blurry controls in QQuickWidget
QQuickWindow::devicePixelRatio doesn't take into account being inside a
QQuickWidget or other render target. This matters if the window dpr differs
from the screen dpr, something we see on wayland.

QQuickWindow has another similarly named method which does, with docs
stating this is the one that should always be used.