Menu: Factor out property assignment
Assigning to a property in a tight loop is inefficient. Better create
some local variables for such purpose.

Note that once set to true, neither of those two properties can reset
back to false. This is probably due to the fact that the code deals with
only visible subset of a ListView's delegates.