Revisit usages of QtQuick.Controls/AbstractButton::icon grouped property
Here are the rules:

1. AbstractButton::icon is a grouped property. It can not be used as a
conditional expression by itself as it would be quite pointless,
because it's always considered truthy like any normal object in js.

2.a. icon.name is a string, which itself can be compared to an empty
string "", or its .length property compared to zero.

2.b. icon.source is a QUrl, or a `url` as QmlEngine likes to call it. It
is an opaque object for JavaScript environment, which provides magical
toString() method that makes it look like a string in some contexts.
Strict equality with real strings will fail, as well as attempts to
access its .length directly. Empty url converts to an empty string.

2.c. KQuickStyleItem accepts both strings and urls as `properties["icon"]` value.

4. Other sub-properties of icon are always present and default to some
values such as 0 size and transparent color.