ScrollView: Factor out subexpression in a way that actually works reliably
Due to the nature of non-deterministic order of signal propagation, one
may not rely on properties to hold factored out common parts of
expressions and rely on them as null-guards or anything like that.

On the contrary, methods defined in pure QML, when used as a part of
binding expression, will transitively bind to objects and properties
that they access while evaluating.

(cherry picked from commit 9e685a63bd2c138af849f2287dfa7a1c08663b29)