ScrollBar: Fix visuals in edge cases of rounding errors
At certain sizes, it was possible to end up with a scrollbar that would
look like it can scroll further down/right when it was actually at the
far end. This could've happened due to automatic rounding -- because
styleitem's value & maximum are integer properties, and loss of
precision -- while dividing and multiplying an already computed
floating number.