Drawer: Fix RTL by copying sizing code from upstream Default style
There was this unused property `horizontal`, which was clearly copied
from there some time ago, and for whatever reason rewritten in anchors.
Disadvantage of using anchors is that they inherit LayoutMirroring by
default, unless explicitly requested not to. But Drawer::edge property
is not mirrored at all. Thus, replacing custom layout code with the
upstream's also fixes visual issues for right-to-left environments.