Don't attempt to sync colors when application is shutting down
This prevents the odd crash when the application is closing (and qApp
could possibly be null), but QQuickItem::visibleChanged is still calling
but technically not in the destructor - so
PlasmaDesktopTheme::syncColors is still being called. This drops down to
KConfigWidgets, who attempts to use a QCoreApplication that doesn't
exist.

There's no point in syncing colors in a closing application, so let's
just early exit here.