KQuickStyleItem: Update visuals after random focusReason changes
It may happen, especially with custom keyboard navigation logic, that
focusReason is being set separately after control has gained an active
focus, so we need to listen for those changes as well.

Additionally, with the new property notification binding we don't need
to react to FocusIn/Out events in event filter.

BUG: 469215