ToolTip: Do not rely on Layout
On Qt >= 6.5, usage of the Layout results in binding loops (as the
Layout ends up reapplying size hints multiple times, leading to a
spurious binding loop warning on contentWidth).
Given that there was already a comment advising to rather not use
Layouts (due to their bad interaction with child items using anchors),
we sidestep the issue by forcing the word-wap behaviour via
logic in onLineLaidOut.

Bug: 472872