ProgressBar: Rewrite for the greater good
Get rid of nasty math! Embrace simplicity and perfection of `position`!

Now the range of possible bridged values is totally predictable, safe
from overflows, and precise enough to express fractions of progress
even on the biggest available monitors.

For an ultimate effect, requires a patched Breeze with fixed RTL rendering.