Slider: Drop vertical centering
Overriding anchors of controls' contentItem and background is
technically questionable and prone to breakage. Besides, it ruins the
idea of overriding padding values.

Sliders were not meant to be resized in the dimension orthogonal to
their direction, so this kind of adjustment should not be needed in
well-designed layouts. In practice, StyleItem will be stretched to fill
control (minus padding/insets), and from there is it up to a style to
render it centered or not. In my testing, Oxygen based styles (including
Breeze) stick their slider to the top, Fusion centers them, while
MS Windows 9x stretches slider handle all across.