ScrollBar: Fix size being initially stuck on Qt 6
Due to a bug in Qt, explicit binding to ScrollBar::size is required to
help it notify the binding to reevaluate. According to my and Noah's
investigation, size can indeed change in a way that affects getter of
visualSize, but only signal sizeChanged will be dispatched, thus leaving
visualSize hanging in a limbo state when querying it manually would
result in a new value but change listeners would never know that.

BUG: 473241
See also: https://codereview.qt-project.org/c/qt/qtdeclarative/+/501319