HorizontalHeaderView: Read `sort` role to determine if we should show sorting
Since the header lacks API to set this, we need to come up with a
different solution. This was the most minimal that I could get it.
Behaviour now is, if there's no sort role or it explicitly returns
undefined, we don't show any sort indicator. If there is, it is assumed
to be `Qt::SortOrder` and we use it to show either the up or down arrow.