Make handling of ItemDelegate and subtypes' indicator space consistent
When there is an indicator, we should make sure to reserve space for it
so that applications don't have to deal with it when overriding the
contentItem. Additionally, we want to make sure that the indicator is
placed on the left side of the delegate as that matches with what we
have been doing so far.

Unfortunately, the only way I could find to reserve space for the
indicator is by overriding left/right padding, which means this will
break if it is explicitly set by the application. I don't really know a
better solution however, as anything inside the contentItem breaks as
soon as the contentItem is replaced.