delegates: Enable tooltips by default when text is truncated
This is behavior that used to be the default with BasicListItem and
which was ported over to the new Kirigami content items. However, since
that limits it to the content item, the tooltip will not display when
the content item is too small. Additionally, this means tooltips only
show up when using these content items.

Instead, we can enable the tooltip on the delegates directly, which
means the entire delegate becomes the trigger surface. Since these are
just setting defaults for the tooltip properties, it means they can be
easily overridden when tooltips should be disabled or when it should
display something different.