Slider: Fix slider scrolling direction
We always want the slider to increase its value when scrolling
up or right. This corresponds to positive values on the Y axis
and negative values on the X axis.

This is independent of the inverted scrolling setting.

Note that the Qt documentation on WheelEvent is dubious:
On https://doc.qt.io/qt-5/qml-qtquick-wheelevent.html, the
`angleDelta` section suggest that positive is up/right,
but the `inverted` paragraph mentions the opposite.