Don't set a background color for normal unhovered list items
Currently the background styling we apply to ItemDelegate and others
uses the Kirigami theme's standard view background color for the
backgrounds of unhovered list items not using the alternate background
color. But this is unnecessary; when the list item is in a view that
itself uses the view background color, then the list item's own
background color simply matches that of the view it lives in. And when
the view intentionally does not use that background color, the result is
visual brokenness.

We can fix this by simply not setting a background color for the normal
unhovered state of the background.

BUG: 477792