HeaderView: Paint end item in free space at the end
The end item does not correspond to the last column/row but the the
space of the view where there's no data, this also shows in the
separator being not painted.
Matches what widgets does and has a major UX improvement: The
click target for the resizing the last column is now visible.
BUG:478124
FIXED-IN:6.0