styleitem: Base render node size on image
When using fractional scaling, we may end up with an image that is
slightly different in size from the bounds of our item, since the image
will always be in integer size but the item size can be fractional. This
causes it to be rendered at a slightly different size from the texture
size, which triggers interpolation and leads to glitches in the rendered
result.

To avoid all that, rather than using our item's bounds for the node
size, use the image size converted back to logical pixels as the node
size. This may actually end up with us using a fractional node size, but
ensures that when rendered we use the exact image dimensions.

CCBUG: 464069