HorizontalHeaderView: Fix incorrect highlighting when no selection model set
If there's no selection model for the header, we try to use the
selection model from syncView. However, the header may be using a
completely different model than the sync view. In such cases, using the
sync view's selection model will lead to odd results as the row/column
mappings may not be correct. So don't try to use it and just don't do
any highlighting in that case.