Menu: Use StyleItem as a property
The `count` property of the QQC2.Menu component shows the number of
visual children. StyleItem is a visual element, so it will affect this
property. This behavior can lead to an error when, for example, the
user does not want to show an empty menu when accessing the "count"
property.