Completely disable animations when animations are globally disabled
Right now we rely on this happening implicitly when the relevant
duration values become 0. But an animation with a 0 duration does not
work as expected with certain types of QtQuick Animators. In addition,
even if it does work, a 0-duration animation is inefficient, doing
extra work for no reason. Let's disable the animations entirely when
we know the user has globally disabled animations.