TreeViewDelegate: Fix non-observable modelIndex property getting stuck
Although not quite consistent, the ItemBranchIndicators item could
randomly get stuck without any usable modelIndex being ever set. It was
caused by the `readonly property var modelIndex` not emitting any
change signal, which is turn was caused by `treeView.index(row, column)`
expression being non-observable AND evaluating before a model is set.

This whole fix is quite bad, and I'd rather use an original QModelIndex
instead having to reconstruct it from bits and manually subscribing to
events which are potentially affecting it; but TableView could only set
an `index` required property which an integer that seems to identical
to the `row` property.