Delegates: Use inset to control spacing around hover/selection highlight
Whether or not to use the bordered style now depends on how much inset
there is, if there is none we use the "table" style and if there is some
we use the "list" style. This allows applications control over which
style they want to use for the delegate, without needing new API.

By default we still use the non-border style when we detect being inside
a table, so this should not result in any visual changes.