ScrollBar: Fix custom scrolling behavior
In order to add our custom mouse event handling behavior to the
QQuickScrollBar, the only way in Qt 6 is to put the MouseArea in front
of it, i.e. with a higher stacking z index, and only handle a subset of
events which we are going to fully process ourselves, while rejecting
the rest to let them propagate further down to the scroll bar itself.

BUG: 488092