Use Qt text rendering throughout
It is known that native rendering performs badly with any form of
scaling.

This was replaced with a static check that chooses one renderer at
startup. The check does not work on Wayland that has per-window rather
than per-screen scaling. Even if it did work it would not support
hotplugging.

Changing this global static at runtime does not change any existing
labels. The alternative is setting a render type in every label in every
control. This used to be the case, but given client code can also have
text does not scale well.

For QtQuick the Qt renderer is more performant, handles transformations
better and avoids this issue. Given the results look basically
identical, we can simplify the existing code.