Use Qt text rendering when high DPI scaling
It is known that native rendering performs badly with scaling and an
existing workaround is in place.

The current check does not work on Wayland that has per-window rather
than per-screen scaling. Given Qt changes hinting preferences when any
scaling is used anyway, we may as well commit to using the non-native
rendering throughout.

For QtQuick the Qt renderer is more performant, handles transformations
better and avoids this issue. Given the results look basically
identical, we can simplify the existing code.

BUG: 479891