set Accessible.name on Mnemonic-having labels
QQC2 desktop style automatically sets mnemonics for many types of
controls. This works by setting the text of an inner component to
the rich text format reflecting the alt-(un)pressed state of the label.
The Accessible.names of these controls, however, are implicitly tied to
the text, and so they are updated to the version with rich text
annotation as well. This is without immediate bad consequences, as it is
hard to get orca in a situation where it will print out the label while
the user presses alt. But it is pointless busywork for orca, unnecessary
dbus chatter, makes debugging a11y issues with event logs more annoying,
is not what QtWidgets or Gtk apps do, and if a user somehow manages to
achieve a readout, they will hear a version with <u> markup sprinkled
in.

Instead, set the Accessible.name to the plain text version explicitly.