set Accessible.name for controls with mnemonics
For controls that have or may have an internal label,
Kirigami.MnemonicData adds Mnemonic marking and only shows them while
the user is holding Alt. The control's text property that feeds the
default assignment of Accessibility.name is not affected. However,
if a mnemonic is supplied using "&" in the text, this is also left
unchanged, and screen readers will read it out.

Instead set Accessible.name explicitly to the plain text version that
MnemonicData supplies.