Update to the default styling from Kirigami
Update to the default styling from Kirigami. This centers the dialog in the window if able and uses the same header.

This come out of discussions on deprecating `Kirigami.Dialog` (frameworks/kirigami!1745) which would mean pushing some of it's styling down into the base (I think this is desirable anyway).