ComboBox: Disallow opening empty comboboxes
Don't let the user open a combobox which has zero items in the
popup. Unfortunately there isn't dedicated api for that and the
popup has to be dinamically assigned or unassigned, but it
works well.

the arrow won't yet change its look, as it needs adjustment in Breeze

together with https://invent.kde.org/plasma/breeze/-/merge_requests/554

BUG:507183