Drop outdated workaround making scrollbars off by a frame
According to the linked ticket that bug was fixed 3 years ago and is
picked to the Qt6.2 branch.

The effect of this workaround is that the scollbar would appear one
event later, which given models populate in a polish step is almost
always one frame later. This makes model views look janky on load with
one frame without a scrollbar then it appearing.