DefaultListItemBackground: Introduce an "inactive highlight" effect
There's a state that's often not handled: an item that's the current or
highlighted item in a list or grid view, but that doesn't have focus
right now. This state is useful to communicate for purposes of keyboard
navigation, so it's obvious which item will receive focus when the list
or grid as a whole is focused.

We don't have a dedicated color for this, so I chose one that got us a
grayscale color since a lightened version of a color isn't ideal for
accessibility.

CCBUG: 508850