Equalize top, left, and right paddings for ItemDelegate + subclasses
This ensures perfect equal padding between items and view edges for both
list and grid items, without increasing the spacing between them, or
using side padding inconsistent with other items.

To achieve this, we use a workaround for the lack of
https://bugreports.qt.io/browse/QTBUG-127729. It's not *thaaaat* bad,
and the only user-visible side effects are 2px taller highlight effects,
and pointer hitboxes being shifted upwards by 2px.

BUG: 487653
FIXED-IN: 6.20